Create Local_File_sharing.py
This is a website that runs locally on a server or pc or phone if set up right.. encryption: a method that I use in the code is AES-256... this website uses file upload and sharing... user can choose to upload their files and it will be hosted by the server-side wherever the code is being ran.. having a simple ui showing connected users at live time it. 

Please remember to change the username and password before the server is locally on the network.